pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "levanter"
version = "1.2"
authors = [
    { name = "David Hall", email = "dlwh@cs.stanford.edu" },
    { name = "Ivan Zhou", email = "ivanz@stanford.edu" },
]
description = "Scalable Training for Foundation Models with Named Tensors and JAX"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
]
dependencies = [
    "haliax>=1.4.dev307",
    "equinox>=0.11.4",
    "jaxtyping>=0.2.20",
    "tokenizers>=0.15.2",
    "transformers>=4.41.2",
    "optax>=0.1.9",
    "wandb>=0.17.8",
    "scipy<=1.12.0",
    "draccus>=0.8.0",
    "pyarrow>=11.0.0",
    "zstandard>=0.20.0",
    "datasets>=2.18,<4.0",
    "gcsfs>=2024.2,<2024.10",
    "braceexpand>=0.1.7",
    "jmp>=0.0.3",
    "fsspec[http]>=2024.2,<2024.10",
    "tensorstore>=0.1.65",
    "pytimeparse>=1.1.8",
    "humanfriendly==10.0",
    "safetensors[numpy]~=0.4.2",
    "matplotlib>=3.7.0",
    "tblib>=1.7.0,<4.0.0",
    "dataclasses-json~=0.6.4",
    "ray[default]==2.34.0",
    "pydantic<3",
    "rich~=13.0",
    "filelock~=3.13",
    #    "ai2-olmo",
    "async-lru~=2.0",
    "tqdm-loggable>=0.2",
    "deepdiff"
]

[project.urls]
"Homepage" = "https://github.com/stanford-crfm/levanter"
"Bug Tracker" = "https://github.com/stanford-crfm/levanter/issues"

[tool.black]
line-length = 119
target-version = ["py310"]
preview = true

[tool.isort]
profile = "black"
multi_line_output = 3
lines_after_imports = 2
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 119
src_paths = ["src", "tests"]
known_haliax = ["haliax"]
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "HALIAX",
    "FIRSTPARTY",
    "LOCALFOLDER",
]

[tool.mypy]
python_version = "3.10"
mypy_path = ["src"]

[tool.mypy-haliax.core]
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "entry: marks tests as entry point tests (deselect with '-m \"not entry\"')",
    "ray: marks tests that require Ray (deselect with '-m \"not ray\"')",
]

[project.optional-dependencies]
test = [
    "pytest",
    "flake8",
    "soundfile",
    "librosa",
    "pytest-forked",
    "pytest-asyncio",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["levanter", "levanter.*"]
